Move polling to DashboardPage
